Prevent overflow when dealing with datatype count. This patch fixes #7147 by preventing overflow when multiplying the count and the blocklen. The count reflects MPI count and is therefore bound to the size of an int (it is an uint32_t) while the blocklen can be merged together to represent the largest contiguous memory layout and it is therefore promoted to a size_t.

+++ opal/datatype/opal_datatype_copy.h
@@ -128,7 +128,7 @@ static inline int32_t _copy_content_same_ddt( const opal_datatype_t* datatype, i
     DO_DEBUG( opal_output( 0, "_copy_content_same_ddt( %p, %d, dst %p, src %p )\n",
                            (void*)datatype, count, (void*)destination_base, (void*)source_base ); );
 
-    iov_len_local = count * datatype->size;
+    iov_len_local = (size_t)count * datatype->size;
 
     /* If we have to copy a contiguous datatype then simply
      * do a MEM_OP.
